clog: add new logging methods to publish a set of public logs
The O community has shown interest in B introspection where some logging
can be published to a public facing Loki instance that can be used to
debug the selection algorithm. To that end, this change adds a few
methods to the clog package to publish logs with a specific tag
([PublicLogs]) and removes any contextual info that should not be
published.

To log specific lines to the public facing Loki instance, one can do the
following as an example:
 - clog.PublicInfof(ctx, "Updating O ID=%v, reason=%v", "fooID", "bar")

The resulting log will have specific contextual info like manifestID,
orchSessionID, etc pre-pended along with a [PublicLogs] tag.